PostDetailViewをインタフェースと実装に分ける
